pruebas unitarias a los metodos Crud de JPARepository

+++ src/main/java/com/example/demo/repository/PartidaRepository.java
@@ -1,9 +1,11 @@
 package com.example.demo.repository;
 
+import com.example.demo.modelo.Mensaje;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.modelo.Partida;
 
-public interface PartidaRepository extends JpaRepository<Partida,Integer>{
-    
+import java.util.List;
+
+public interface PartidaRepository extends JpaRepository<Partida,Long>{
 }
